==> 1B_peekbank_item_vis.Rmd <==
---
title: "Peekbank Item Visualization"
author: "Martin"
date: "1/19/2021"
output: html_document
---

Visualize item-specific trajectories

```{r}
library(peekbankr)
library(tidyverse)
library(lme4)
library(lmerTest)
library(tictoc)
library(langcog)
library(here)

figure_path <- here("figures")

t_range <- c(-1000,3000)
knitr::opts_chunk$set(cache = TRUE, warn = FALSE, warning = FALSE, message = FALSE,cache.lazy = FALSE)

load(file = "cached_intermediates/1_aoi_data_joined.Rds")
dataset_name_mapping <- read_csv(here("aux_data","dataset_name_mapping.csv"))

```


Time series - Item

```{r}
t_range <- c(-1000,3000)
aoi_data_joined <- aoi_data_joined |>
  left_join(dataset_name_mapping) |>
  filter(t_norm > t_range[1],
         t_norm < t_range[2])


means_items <- aoi_data_joined |>
  left_join(dataset_name_mapping) |>
  #filter(age > 12, age <= 60) |>
  #mutate(age_binned = cut(age, seq(0,60,12))) |>
  #group_by(t_norm, dataset_name, age_binned, stimulus_novelty) |>
  group_by(t_norm, dataset_name, apa_cite, target_label) |>
  summarise(n = sum(aoi %in% c("target","distractor"), na.rm = TRUE), 
            p = sum(aoi == "target", na.rm = TRUE),
            prop_looking = p / n, 
            ci_lower = binom::binom.confint(p, n, method = "bayes")$lower,
            ci_upper = binom::binom.confint(p, n, method = "bayes")$upper) 

average_across_items <- means_items |>
  group_by(t_norm, dataset_name) |>
  summarize(
    N=n(),
    avg_prop_looking=mean(prop_looking,na.rm=T),
    sd = sd(prop_looking,na.rm=T),
    ci=qt(0.975, N-1)*sd/sqrt(N),
    ci_lower=avg_prop_looking-ci,
    ci_upper=avg_prop_looking+ci
  )

ggplot(filter(means_items, n>=12 & dataset_name!="pomper_dimy"), aes(x = t_norm, y = prop_looking,color=target_label)) + 
  geom_line(alpha=0.5) + 
  # geom_ribbon(aes(ymin = ci_lower, ymax = ci_upper, 
  #                 fill = dataset_name), alpha = .5) +
  # geom_smooth(data=average_across_items,
  #             aes(y=avg_prop_looking,ymin = ci_lower, ymax = ci_upper),color="black",stat="identity")+
  geom_smooth(color="black",se=F,method="gam")+
  geom_hline(yintercept = .5, lty = 2) + 
  geom_vline(xintercept = 0, lty = "solid") +
  ylab("Proportion Target Looking") +
  xlab("Time (msec)") +
  theme_classic() +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  facet_wrap(~apa_cite,nrow=5)+
  theme(legend.position="none")+
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=16))+
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5),strip.text.x = element_text(size = 9))+
  scale_x_continuous(breaks=seq(-500,3000,500))+
  scale_y_continuous(breaks=c(0,0.25,0.5,0.75,1),limits=c(0,1))
  
  
ggsave(here(figure_path,"peekbank_item_vis.png"),width=10,height=6,dpi=600)
```

# Average trial-level data

Goal here is to average looking performance for each trial across a critical window (t_min and t_max).
We also set a threshold for how much looking data must be included in order for the trial to merit inclusion.

```{r}
#### PARAMETERS TO SET ####
#critical window dimensions consistent with measure development
t_min <- 500
t_max <- 4000
#proportion missing trials threshold (any trial in which over half of the critical window missing is looking data is excluded )
max_prop_missing <- 0.5
#age bin size (number of months per bin)
age_bin_size <- 6
#add baseline window for computing baseline-corrected means
baseline_window <- c(-2000,0)
#minimum baseline duration in ms
min_baseline <- 500


by_trial_means <- aoi_data_joined %>%
  #restrict to english datasets (this is just because there are so few non-English datasets atm)
  filter(!(dataset_name %in% c("casillas_tseltal_2015", 
                               "byers-heinlein_2017", 
                               "hurtado_2008", 
                               "kartushina_2019",
                               "weisleder_stl",
                               "xsectional_2007",
                               "pomper_dimy"))) |>
  #restrict age range
  filter(age > 6, age <= 60) %>%
  # familiar target items only %>%
  filter(stimulus_novelty == "familiar") %>%
  #window of analysis
  filter(t_norm >= t_min, t_norm <= t_max) %>%
  #bin ages (can adjust size of age bins here)
  mutate(age_binned = cut(age, seq(12,60,age_bin_size))) %>%
  group_by(dataset_name,subject_id, trial_id, target_label, 
           age, age_binned) %>%
  summarise(prop_target_looking = sum(aoi == "target", na.rm = TRUE) / 
              (sum(aoi == "target", na.rm=TRUE) + 
                 sum(aoi=="distractor", na.rm=TRUE)),
            prop_missing = mean(aoi %in% c("missing","other"), na.rm = TRUE)) %>%
  #remove trials with insufficient looking to target or distractor
  filter(prop_missing<=max_prop_missing)

#compute baseline looking (for baseline-corrected means)
by_trial_baseline <- aoi_data_joined %>%
  #restrict to english datasets (this is just because there are so few non-English datasets atm)
  filter(!(dataset_name %in% c("casillas_tseltal_2015", 
                               "byers-heinlein_2017", 
                               "hurtado_2008", 
                               "kartushina_2019",
                               "weisleder_stl",
                               "xsectional_2007",
                               "pomper_dimy"))) |>
  #restrict age range
  filter(age > 12, age <= 60) %>%
  # familiar target items only %>%
  filter(stimulus_novelty == "familiar") %>%
  #window of analysis
  filter(t_norm >= baseline_window[1], t_norm <= baseline_window[2]) %>%
  #bin ages (can adjust size of age bins here)
  mutate(age_binned = cut(age, seq(12,60,age_bin_size))) %>%
  group_by(dataset_name,subject_id, trial_id, target_label, 
           age, age_binned) %>%
  summarise(
   baseline_n=n(),
   baseline_ms=baseline_n*25,
    baseline_looking = sum(aoi == "target", na.rm = TRUE) / 
              (sum(aoi == "target", na.rm=TRUE) + 
                 sum(aoi=="distractor", na.rm=TRUE)),
            prop_baseline_missing = mean(aoi %in% c("missing","other"), na.rm = TRUE)) %>%
  #remove trials with insufficient looking to target or distractor
  filter(prop_baseline_missing<=max_prop_missing& baseline_ms>=500)

#combine
by_trial_target_means <- by_trial_means %>%
  left_join(by_trial_baseline) %>%
  mutate(corrected_target_looking=prop_target_looking-baseline_looking)

```

# Average within subjects, by-item and by-dataset

One could consider excluding participants based on the number of trials a participant contributes overall here.

```{r}
by_subj_item_means <- by_trial_target_means %>%
  group_by(dataset_name,subject_id, target_label, 
           age, age_binned) %>%
  summarise(
    trial_num=n(),
    avg_target_looking = mean(prop_target_looking,na.rm=TRUE),
    avg_corrected_target_looking=mean(corrected_target_looking,na.rm=TRUE)
  )
```

# Average across subjects - by item, dataset and age bin

```{r}
by_item_means <- by_subj_item_means %>%
  group_by(dataset_name, target_label,age_binned) %>%
  summarise(
    subj_n=n(),
    target_looking = mean(avg_target_looking,na.rm=TRUE),
    corrected_looking = mean(avg_corrected_target_looking,na.rm=TRUE)
  )
```

# Mega Plot of items (collapsing across age & dataset)

Note that because this collapses across age, some items might be showing higher preformance because we only have observations for older children (who will typically do better overall).

```{r}
ggplot(by_item_means,aes(reorder(target_label,target_looking,mean),target_looking,color=target_label))+
  geom_boxplot()+
  #geom_point()+
  theme(legend.position="none")+
  theme(axis.text.x=element_text(angle=90,size=4,vjust=0.5))+
  xlab("Target Label")+
  ylab("Proportion Target Looking")

ggplot(filter(by_item_means,!is.na(corrected_looking)),aes(reorder(target_label,corrected_looking,mean),corrected_looking,color=target_label))+
  geom_boxplot()+
  #geom_point()+
  theme(legend.position="none")+
  theme(axis.text.x=element_text(angle=90,size=4,vjust=0.5))+
  xlab("Target Label")+
  ylab("Proportion Corrected Target Looking")
```

```{r}
by_item_means_across_dataset <- by_item_means %>%
  group_by(age_binned,target_label) %>%
  summarize(
    dataset_num=n(),
    prop_target_looking=mean(target_looking,na.rm=TRUE),
    corrected_target_looking=mean(corrected_looking,na.rm=TRUE)
  )

by_item_means <- by_item_means %>%
  left_join(select(by_item_means_across_dataset,-prop_target_looking,-corrected_target_looking))

ggplot(filter(by_item_means,dataset_num>0),aes(reorder(target_label,target_looking,mean),target_looking,color=reorder(target_label,target_looking,mean)))+
  geom_boxplot()+
  geom_hline(yintercept=0.5,linetype="dashed")+
  xlab("Target Label")+
  ylab("Proportion Target Looking")+
  scale_color_viridis_d()+
  theme_cowplot(font_size=20)+
  theme(legend.position="none")+
  theme(axis.text.x=element_text(angle=90,size=6,vjust=0.5))
ggsave(here(figure_path,"peekbank_item_accuracies.png"),width=10,height=6,dpi=600)
```
